split out result components into separate files as clean-up, changed criteriaevaluationresult back to regular enum for cleaner state
