+++ teachertool/src/components/styling/EvalResultDisplay.module.scss
@@ -35,3 +35,10 @@
     justify-content: space-between;
     // height: 2.625rem;
 }
+
+.result-dropdown {
+    font-size: 9px;
+    height: 1.5rem;
+    margin: 0.5rem 0;
+    color: black;
+}
