+++ teachertool/src/types/criteria.ts
@@ -31,10 +31,10 @@ export interface CriteriaParameterValue {
 
 // Possible results from evaluating a criteria instance.
 export enum CriteriaEvaluationResult {
-    Pass = "Looks Good!",
-    Fail = "Needs Work",
-    CompleteWithNoResult = "Not evaluated",
-    InProgress = "Evaluating...",
+    Pass,
+    Fail,
+    CompleteWithNoResult,
+    InProgress,
 }
 
 export interface CriteriaResult {
